requirements.txt
# AI Content Generation Suite - Consolidated Requirements
# This file contains all dependencies for the entire project

# ===== Core Dependencies =====
python-dotenv>=1.0.0      # Environment variable management
requests>=2.31.0          # HTTP library
typing-extensions>=4.0.0  # Type hints backport for older Python versions
pyyaml>=6.0               # YAML configuration parsing
pathlib2>=2.3.7           # Path manipulation library
argparse>=1.4.0           # Command-line argument parsing

# ===== AI/ML Service Clients =====
# FAL AI
fal-client>=0.4.0         # FAL AI API client

# Replicate
replicate>=0.15.0         # Replicate API client

# OpenAI
openai>=1.0.0,<2.0.0      # OpenAI API client (for OpenRouter integration)

# Google Cloud AI
google-cloud-aiplatform>=1.20.0  # Vertex AI client (relaxed for Python 3.8)
google-cloud-storage>=2.0.0      # Google Cloud Storage (relaxed for Python 3.8)
google-auth>=2.0.0                # Google authentication (relaxed for Python 3.8)
google-generativeai>=0.2.0      # Google Generative AI (used by legacy code)
google-genai>=1.0.0             # Google GenAI SDK (new unified SDK)

# Text-to-Speech
elevenlabs>=1.0.0         # ElevenLabs TTS API

# ===== Media Processing =====
Pillow>=10.0.0            # Image processing library
moviepy>=1.0.3            # Video editing library
ffmpeg-python>=0.2.0      # FFmpeg Python bindings

# ===== HTTP and Async =====
aiohttp>=3.8.0            # Async HTTP client/server
httpx>=0.25.0             # Modern HTTP client

# ===== Development Tools =====
# Jupyter/Notebooks
jupyter>=1.0.0            # Jupyter metapackage
ipython>=8.0.0            # Enhanced Python shell
notebook>=7.0.0           # Jupyter notebook

# Testing
pytest>=7.0.0             # Testing framework
pytest-asyncio>=0.21.0    # Async test support

# ===== Additional Tools =====
# mcp>=1.0.0                # MCP server support (not available on PyPI)

# ===== Optional/Conditional Dependencies =====
